Move the printing of recordings directly to `testing/requestErrors`
